requirements.txt
apturl==0.5.2
async-generator==1.10
attrs==22.1.0
awscli==1.22.34
bcrypt==3.2.0
blinker==1.4
botocore==1.23.34
Brlapi==0.8.3
cachetools==5.0.0
Cerberus==1.3.4
certifi==2022.9.24
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
docutils==0.17.1
duplicity==0.8.21
exceptiongroup==1.0.0
fasteners==0.14.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
h11==0.14.0
HTMLTestRunner-Python3==0.8.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
jeepney==0.7.1
jmespath==0.10.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.12.4
ptyprocess==0.7.0
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pyspark==3.3.1
pytest==7.2.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
selenium==4.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
systemd-python==234
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
ttystatus==0.38
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
